move things around a little
