+++ function/builtin/aggregate/aggregate.go
