+++ function/builtin/forecast/anomaly_function.go
