+++ function/builtin/forecast/anomaly_test.go
