+++ function/builtin/forecast/drop_function.go
