+++ function/builtin/forecast/utility.go
