+++ function/builtin/transform/special.go
