+++ main/web/static/script.js
