+++ main/web/static/style_embed.css
