+++ metric_metadata/cassandra/schema/schema_test.cql
